backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.12
pydantic==2.5.1
PyPDF2==3.0.1
python-docx==0.8.11
supabase==2.6.0
postgrest>=0.14,<0.17.0
storage3>=0.5.3,<0.8.0
supafunc>=0.3.1,<0.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
